Feat: remove unnecessary control value accessors

+++ src/app/core/pages/client/pages/menu/layout/menu.component.ts
@@ -1,23 +1,21 @@
 import type { OnDestroy, OnInit } from "@angular/core";
 import { ChangeDetectionStrategy, Component } from "@angular/core";
-import { FormControl } from "@ngneat/reactive-forms";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
-import { lastValueFrom, map, switchMap, tap } from "rxjs";
+import { filter, lastValueFrom, map, ReplaySubject, switchMap, take, tap } from "rxjs";
 import { CATEGORY_ID, ORDER_ID, PLACE_ID } from "src/app/shared/constants";
 import { CLIENT_ROUTES } from "src/app/shared/constants";
 import { BreadcrumbsService } from "src/app/shared/modules/breadcrumbs";
 import { RouterService } from "src/app/shared/modules/router";
 
-import type { ProductEntity } from "../../../../../../../graphql";
 import { ProductToOrderStatusEnum } from "../../../../../../../graphql";
 import { ActionsService } from "../../../../../../features/app";
 import { OrdersService } from "../../../../../../features/orders";
-import type { IEmit } from "../../../../../../features/products";
 import { ProductDialogComponent } from "../../../../../../features/products";
-import type { DeepPartial } from "../../../../../../shared/interfaces";
+import type { IProductToSelect } from "../../../../../../features/products/ui/products-select/interfaces";
 import { DialogService } from "../../../../../../shared/ui/dialog";
 import { MENU_PAGE_I18N } from "../constants";
 import { MenuPageCategoriesGQL, MenuPageOrderGQL, MenuPageProductsGQL } from "../graphql/menu-page";
+import type { IProductChanged } from "../interfaces";
 
 @UntilDestroy()
 @Component({
@@ -27,46 +25,56 @@ import { MenuPageCategoriesGQL, MenuPageOrderGQL, MenuPageProductsGQL } from "..
 	changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class MenuComponent implements OnInit, OnDestroy {
-	readonly menuPagei18n = MENU_PAGE_I18N;
-
+	readonly menuPageI18n = MENU_PAGE_I18N;
 	private readonly _menuPageCategoriesQuery = this._menuPageCategoriesGQL.watch();
+	private readonly _menuPageProductsQuery = this._menuPageProductsGQL.watch();
+	private readonly _menuPageOrderQuery = this._menuPageOrderGQL.watch();
+	private readonly _selectedCategorySubject = new ReplaySubject<string>();
+	readonly selectedCategory$ = this._selectedCategorySubject.asObservable();
 	readonly categories$ = this._menuPageCategoriesQuery.valueChanges.pipe(
 		map((result) => result.data.categories.data),
-		tap((categories) => {
-			if (!categories || !categories[0] || this.categoryControl.value) {
+		tap(async (categories) => {
+			const { categoryId, placeId } = this._routerService.getParams();
+
+			if (!categories || !categories[0] || categoryId) {
 				return;
 			}
 
-			this.categoryControl.setValue(categories[0].id);
+			await this._routerService.navigateByUrl(
+				CLIENT_ROUTES.CATEGORY.absolutePath.replace(PLACE_ID, placeId).replace(CATEGORY_ID, categories[0].id)
+			);
 		})
 	);
 
-	private readonly _menuPageOrderQuery = this._menuPageOrderGQL.watch();
-	private readonly _menuPageProductsQuery = this._menuPageProductsGQL.watch();
-	readonly products$ = this._menuPageProductsQuery.valueChanges.pipe(
-		map((result) => result.data.products.data),
-		switchMap((products) =>
-			this._menuPageOrderQuery.valueChanges.pipe(
-				map((result) => result.data.order),
-				map((order) =>
-					products?.map((product) => ({
-						...product,
-						productsToOrders: (order.productsToOrders || [])
-							.filter(
-								(productToOrder) =>
-									productToOrder.product.id === product.id && productToOrder.status === ProductToOrderStatusEnum.Added
-							)
-							.map((productToOrder) => ({
-								...productToOrder,
-								attributes: productToOrder.attributes || []
+	readonly products$ = this.selectedCategory$.pipe(
+		take(1),
+		switchMap(() =>
+			this._menuPageProductsQuery.valueChanges.pipe(
+				map((result) => result.data.products.data),
+				switchMap((products) =>
+					this._menuPageOrderQuery.valueChanges.pipe(
+						map((result) => result.data.order),
+						map((order) =>
+							(products || []).map((product) => ({
+								...product,
+								productsToOrders: (order.productsToOrders || [])
+									.filter(
+										(productToOrder) =>
+											productToOrder.product.id === product.id &&
+											productToOrder.status === ProductToOrderStatusEnum.Added
+									)
+									.map((productToOrder) => ({
+										...productToOrder,
+										attributes: productToOrder.attributes || []
+									}))
 							}))
-					}))
+						)
+					)
 				)
 			)
 		)
 	);
 
-	readonly categoryControl = new FormControl<string>();
 	constructor(
 		private readonly _menuPageCategoriesGQL: MenuPageCategoriesGQL,
 		private readonly _menuPageProductsGQL: MenuPageProductsGQL,
@@ -85,47 +93,46 @@ export class MenuComponent implements OnInit, OnDestroy {
 			return;
 		}
 
+		this._routerService
+			.selectParams(CATEGORY_ID.slice(1))
+			.pipe(untilDestroyed(this))
+			.subscribe(async (categoryId) => {
+				if (!categoryId) {
+					return;
+				}
+
+				this._selectedCategorySubject.next(categoryId);
+
+				await this._menuPageProductsQuery.setVariables({
+					filtersArgs: [{ key: "category.id", operator: "=", value: categoryId }]
+				});
+			});
+
 		this._breadcrumbsService.setBreadcrumb({
 			routerLink: CLIENT_ROUTES.CREATE_ORDER.absolutePath.replace(PLACE_ID, placeId)
 		});
+
+		const orderId = await lastValueFrom(
+			this._ordersService.activeOrderId$.pipe(
+				filter((orderId) => Boolean(orderId)),
+				take(1)
+			)
+		);
+
+		if (!orderId) {
+			return;
+		}
+
+		await this._menuPageOrderQuery.refetch({ orderId });
 
 		await this._menuPageCategoriesQuery.setVariables({
 			filtersArgs: [{ key: "place.id", operator: "=", value: placeId }]
 		});
 
-		this._routerService
-			.selectParams(CATEGORY_ID.slice(1))
-			.pipe(untilDestroyed(this))
-			.subscribe(async (categoryId) => {
-				this.categoryControl.setValue(categoryId);
-			});
-
-		this.categoryControl.valueChanges.pipe(untilDestroyed(this)).subscribe(async (categoryId) => {
-			if (!categoryId || categoryId === this._routerService.getParams(CATEGORY_ID.slice(1))) {
-				return;
-			}
-
-			await this._menuPageProductsQuery.setVariables({
-				filtersArgs: [{ key: "category.id", operator: "=", value: categoryId }]
-			});
-
-			await this._routerService.navigateByUrl(
-				CLIENT_ROUTES.CATEGORY.absolutePath.replace(PLACE_ID, placeId).replace(CATEGORY_ID, categoryId)
-			);
-		});
-
-		this._ordersService.activeOrderId$.pipe(untilDestroyed(this)).subscribe(async (orderId) => {
-			if (!orderId) {
-				return;
-			}
-
-			await this._menuPageOrderQuery.refetch({ orderId });
-		});
-
 		this._actionsService.setAction({
 			label: "Подвтердить",
 			func: async () => {
-				const activeOrderId = this._ordersService.getActiveOrderId();
+				const activeOrderId = await lastValueFrom(this._ordersService.activeOrderId$.pipe(take(1)));
 
 				if (!activeOrderId) {
 					return;
@@ -138,7 +145,15 @@ export class MenuComponent implements OnInit, OnDestroy {
 		});
 	}
 
-	async openProductDialog(data: DeepPartial<ProductEntity>) {
+	async setSelectedCategory(categoryId: string) {
+		await this._routerService.navigateByUrl(
+			CLIENT_ROUTES.CATEGORY.absolutePath
+				.replace(PLACE_ID, this._routerService.getParams(PLACE_ID.slice(1)))
+				.replace(CATEGORY_ID, categoryId)
+		);
+	}
+
+	async openProductDialog(data: IProductToSelect) {
 		const result = await lastValueFrom(this._dialogService.open(ProductDialogComponent, { data }).afterClosed$);
 
 		if (!result) {
@@ -148,13 +163,17 @@ export class MenuComponent implements OnInit, OnDestroy {
 		await this.addProductToOrder({ productId: result.product.id, attributesIds: [] });
 	}
 
-	async removeProductFromOrder({ productId, attributesIds }: IEmit) {
-		const activeOrderId = this._ordersService.getActiveOrderId();
+	async removeProductFromOrder({ productId, attributesIds }: IProductChanged) {
+		const activeOrderId = await lastValueFrom(this._ordersService.activeOrderId$.pipe(take(1)));
+
+		if (!activeOrderId) {
+			return;
+		}
 
 		await lastValueFrom(
 			this._ordersService.removeProductFromOrder({
 				productId,
-				orderId: activeOrderId!,
+				orderId: activeOrderId,
 				attrs: attributesIds
 			})
 		);
@@ -162,8 +181,8 @@ export class MenuComponent implements OnInit, OnDestroy {
 		await this._menuPageOrderQuery.refetch();
 	}
 
-	async addProductToOrder({ productId, attributesIds }: IEmit) {
-		const activeOrderId = this._ordersService.getActiveOrderId();
+	async addProductToOrder({ productId, attributesIds }: IProductChanged) {
+		const activeOrderId = await lastValueFrom(this._ordersService.activeOrderId$.pipe(take(1)));
 
 		if (!activeOrderId) {
 			return;
